Rename s/shape-complex/shaper/g
