+++ src/gen-indic-table.py
@@ -106,7 +106,7 @@ print ('#include "hb.hh"')
 print ()
 print ('#ifndef HB_NO_OT_SHAPE')
 print ()
-print ('#include "hb-ot-shape-complex-indic.hh"')
+print ('#include "hb-ot-shaper-indic.hh"')
 print ()
 
 # Shorten values
