+++ src/hb-ot-shaper-khmer-machine.hh
@@ -1,5 +1,5 @@
 
-#line 1 "hb-ot-shape-complex-khmer-machine.rl"
+#line 1 "hb-ot-shaper-khmer-machine.rl"
 /*
  * Copyright © 2011,2012  Google, Inc.
  *
@@ -38,7 +38,7 @@ enum khmer_syllable_type_t {
 };
 
 
-#line 42 "hb-ot-shape-complex-khmer-machine.hh"
+#line 42 "hb-ot-shaper-khmer-machine.hh"
 #define khmer_syllable_machine_ex_C 1u
 #define khmer_syllable_machine_ex_Coeng 14u
 #define khmer_syllable_machine_ex_DOTTEDCIRCLE 12u
@@ -56,7 +56,7 @@ enum khmer_syllable_type_t {
 #define khmer_syllable_machine_ex_ZWNJ 5u
 
 
-#line 60 "hb-ot-shape-complex-khmer-machine.hh"
+#line 60 "hb-ot-shaper-khmer-machine.hh"
 static const unsigned char _khmer_syllable_machine_trans_keys[] = {
 	5u, 26u, 5u, 21u, 5u, 26u, 5u, 21u, 1u, 16u, 5u, 21u, 5u, 26u, 5u, 21u, 
 	5u, 26u, 5u, 21u, 5u, 21u, 5u, 26u, 5u, 21u, 1u, 16u, 5u, 21u, 5u, 26u, 
@@ -239,11 +239,11 @@ static const int khmer_syllable_machine_error = -1;
 static const int khmer_syllable_machine_en_main = 20;
 
 
-#line 43 "hb-ot-shape-complex-khmer-machine.rl"
+#line 43 "hb-ot-shaper-khmer-machine.rl"
 
 
 
-#line 86 "hb-ot-shape-complex-khmer-machine.rl"
+#line 86 "hb-ot-shaper-khmer-machine.rl"
 
 
 #define found_syllable(syllable_type) \
@@ -262,7 +262,7 @@ find_syllables_khmer (hb_buffer_t *buffer)
   int cs;
   hb_glyph_info_t *info = buffer->info;
   
-#line 266 "hb-ot-shape-complex-khmer-machine.hh"
+#line 266 "hb-ot-shaper-khmer-machine.hh"
 	{
 	cs = khmer_syllable_machine_start;
 	ts = 0;
@@ -270,7 +270,7 @@ find_syllables_khmer (hb_buffer_t *buffer)
 	act = 0;
 	}
 
-#line 106 "hb-ot-shape-complex-khmer-machine.rl"
+#line 106 "hb-ot-shaper-khmer-machine.rl"
 
 
   p = 0;
@@ -278,7 +278,7 @@ find_syllables_khmer (hb_buffer_t *buffer)
 
   unsigned int syllable_serial = 1;
   
-#line 282 "hb-ot-shape-complex-khmer-machine.hh"
+#line 282 "hb-ot-shaper-khmer-machine.hh"
 	{
 	int _slen;
 	int _trans;
@@ -292,7 +292,7 @@ _resume:
 #line 1 "NONE"
 	{ts = p;}
 	break;
-#line 296 "hb-ot-shape-complex-khmer-machine.hh"
+#line 296 "hb-ot-shaper-khmer-machine.hh"
 	}
 
 	_keys = _khmer_syllable_machine_trans_keys + (cs<<1);
@@ -315,27 +315,27 @@ _eof_trans:
 	{te = p+1;}
 	break;
 	case 8:
-#line 82 "hb-ot-shape-complex-khmer-machine.rl"
+#line 82 "hb-ot-shaper-khmer-machine.rl"
 	{te = p+1;{ found_syllable (khmer_non_khmer_cluster); }}
 	break;
 	case 10:
-#line 80 "hb-ot-shape-complex-khmer-machine.rl"
+#line 80 "hb-ot-shaper-khmer-machine.rl"
 	{te = p;p--;{ found_syllable (khmer_consonant_syllable); }}
 	break;
 	case 12:
-#line 81 "hb-ot-shape-complex-khmer-machine.rl"
+#line 81 "hb-ot-shaper-khmer-machine.rl"
 	{te = p;p--;{ found_syllable (khmer_broken_cluster); }}
 	break;
 	case 11:
-#line 82 "hb-ot-shape-complex-khmer-machine.rl"
+#line 82 "hb-ot-shaper-khmer-machine.rl"
 	{te = p;p--;{ found_syllable (khmer_non_khmer_cluster); }}
 	break;
 	case 1:
-#line 80 "hb-ot-shape-complex-khmer-machine.rl"
+#line 80 "hb-ot-shaper-khmer-machine.rl"
 	{{p = ((te))-1;}{ found_syllable (khmer_consonant_syllable); }}
 	break;
 	case 5:
-#line 81 "hb-ot-shape-complex-khmer-machine.rl"
+#line 81 "hb-ot-shaper-khmer-machine.rl"
 	{{p = ((te))-1;}{ found_syllable (khmer_broken_cluster); }}
 	break;
 	case 3:
@@ -353,16 +353,16 @@ _eof_trans:
 	case 4:
 #line 1 "NONE"
 	{te = p+1;}
-#line 81 "hb-ot-shape-complex-khmer-machine.rl"
+#line 81 "hb-ot-shaper-khmer-machine.rl"
 	{act = 2;}
 	break;
 	case 9:
 #line 1 "NONE"
 	{te = p+1;}
-#line 82 "hb-ot-shape-complex-khmer-machine.rl"
+#line 82 "hb-ot-shaper-khmer-machine.rl"
 	{act = 3;}
 	break;
-#line 366 "hb-ot-shape-complex-khmer-machine.hh"
+#line 366 "hb-ot-shaper-khmer-machine.hh"
 	}
 
 _again:
@@ -371,7 +371,7 @@ _again:
 #line 1 "NONE"
 	{ts = 0;}
 	break;
-#line 375 "hb-ot-shape-complex-khmer-machine.hh"
+#line 375 "hb-ot-shaper-khmer-machine.hh"
 	}
 
 	if ( ++p != pe )
@@ -387,7 +387,7 @@ _again:
 
 	}
 
-#line 114 "hb-ot-shape-complex-khmer-machine.rl"
+#line 114 "hb-ot-shaper-khmer-machine.rl"
 
 }
