+++ src/hb-ot-shaper-khmer.hh
@@ -29,7 +29,7 @@
 
 #include "hb.hh"
 
-#include "hb-ot-shape-complex-indic.hh"
+#include "hb-ot-shaper-indic.hh"
 
 
 /* buffer var allocations */
