+++ src/hb-ot-shaper-use-table.hh
@@ -42,7 +42,7 @@
 
 #include "hb.hh"
 
-#include "hb-ot-shape-complex-use-machine.hh"
+#include "hb-ot-shaper-use-machine.hh"
 
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-macros"
