+++ src/meson.build
@@ -135,30 +135,30 @@ hb_base_sources = files(
   'hb-ot-os2-unicode-ranges.hh',
   'hb-ot-post-macroman.hh',
   'hb-ot-post-table.hh',
-  'hb-ot-shape-complex-arabic-fallback.hh',
-  'hb-ot-shape-complex-arabic-joining-list.hh',
-  'hb-ot-shape-complex-arabic-table.hh',
-  'hb-ot-shape-complex-arabic-win1256.hh',
-  'hb-ot-shape-complex-arabic.cc',
-  'hb-ot-shape-complex-arabic.hh',
-  'hb-ot-shape-complex-default.cc',
-  'hb-ot-shape-complex-hangul.cc',
-  'hb-ot-shape-complex-hebrew.cc',
-  'hb-ot-shape-complex-indic-table.cc',
-  'hb-ot-shape-complex-indic.cc',
-  'hb-ot-shape-complex-indic.hh',
-  'hb-ot-shape-complex-khmer.cc',
-  'hb-ot-shape-complex-khmer.hh',
-  'hb-ot-shape-complex-myanmar.cc',
-  'hb-ot-shape-complex-myanmar.hh',
-  'hb-ot-shape-complex-syllabic.cc',
-  'hb-ot-shape-complex-syllabic.hh',
-  'hb-ot-shape-complex-thai.cc',
-  'hb-ot-shape-complex-use-table.hh',
-  'hb-ot-shape-complex-use.cc',
-  'hb-ot-shape-complex-vowel-constraints.cc',
-  'hb-ot-shape-complex-vowel-constraints.hh',
-  'hb-ot-shape-complex.hh',
+  'hb-ot-shaper-arabic-fallback.hh',
+  'hb-ot-shaper-arabic-joining-list.hh',
+  'hb-ot-shaper-arabic-table.hh',
+  'hb-ot-shaper-arabic-win1256.hh',
+  'hb-ot-shaper-arabic.cc',
+  'hb-ot-shaper-arabic.hh',
+  'hb-ot-shaper-default.cc',
+  'hb-ot-shaper-hangul.cc',
+  'hb-ot-shaper-hebrew.cc',
+  'hb-ot-shaper-indic-table.cc',
+  'hb-ot-shaper-indic.cc',
+  'hb-ot-shaper-indic.hh',
+  'hb-ot-shaper-khmer.cc',
+  'hb-ot-shaper-khmer.hh',
+  'hb-ot-shaper-myanmar.cc',
+  'hb-ot-shaper-myanmar.hh',
+  'hb-ot-shaper-syllabic.cc',
+  'hb-ot-shaper-syllabic.hh',
+  'hb-ot-shaper-thai.cc',
+  'hb-ot-shaper-use-table.hh',
+  'hb-ot-shaper-use.cc',
+  'hb-ot-shaper-vowel-constraints.cc',
+  'hb-ot-shaper-vowel-constraints.hh',
+  'hb-ot-shaper.hh',
   'hb-ot-shape-fallback.cc',
   'hb-ot-shape-fallback.hh',
   'hb-ot-shape-normalize.cc',
@@ -206,19 +206,19 @@ hb_base_ragel_generated_sources = files(
   'hb-buffer-deserialize-json.hh',
   'hb-buffer-deserialize-text.hh',
   'hb-number-parser.hh',
-  'hb-ot-shape-complex-indic-machine.hh',
-  'hb-ot-shape-complex-khmer-machine.hh',
-  'hb-ot-shape-complex-myanmar-machine.hh',
-  'hb-ot-shape-complex-use-machine.hh',
+  'hb-ot-shaper-indic-machine.hh',
+  'hb-ot-shaper-khmer-machine.hh',
+  'hb-ot-shaper-myanmar-machine.hh',
+  'hb-ot-shaper-use-machine.hh',
 )
 hb_base_ragel_sources = [
   'hb-buffer-deserialize-json.rl',
   'hb-buffer-deserialize-text.rl',
   'hb-number-parser.rl',
-  'hb-ot-shape-complex-indic-machine.rl',
-  'hb-ot-shape-complex-khmer-machine.rl',
-  'hb-ot-shape-complex-myanmar-machine.rl',
-  'hb-ot-shape-complex-use-machine.rl',
+  'hb-ot-shaper-indic-machine.rl',
+  'hb-ot-shaper-khmer-machine.rl',
+  'hb-ot-shaper-myanmar-machine.rl',
+  'hb-ot-shaper-use-machine.rl',
 ]
 
 hb_base_headers = files(
