+++ src/update-unicode-tables.make
@@ -1,29 +1,29 @@
 #!/usr/bin/env -S make -f
 
 all: packtab \
-	hb-ot-shape-complex-arabic-joining-list.hh \
-	hb-ot-shape-complex-arabic-table.hh hb-unicode-emoji-table.hh \
-	hb-ot-shape-complex-indic-table.cc hb-ot-tag-table.hh \
-	hb-ucd-table.hh hb-ot-shape-complex-use-table.hh \
-	hb-ot-shape-complex-vowel-constraints.cc
+	hb-ot-shaper-arabic-joining-list.hh \
+	hb-ot-shaper-arabic-table.hh hb-unicode-emoji-table.hh \
+	hb-ot-shaper-indic-table.cc hb-ot-tag-table.hh \
+	hb-ucd-table.hh hb-ot-shaper-use-table.hh \
+	hb-ot-shaper-vowel-constraints.cc
 
 .PHONY: all clean packtab
 
-hb-ot-shape-complex-arabic-joining-list.hh: gen-arabic-joining-list.py ArabicShaping.txt Scripts.txt
+hb-ot-shaper-arabic-joining-list.hh: gen-arabic-joining-list.py ArabicShaping.txt Scripts.txt
 	./$^ > $@ || ($(RM) $@; false)
-hb-ot-shape-complex-arabic-table.hh: gen-arabic-table.py ArabicShaping.txt UnicodeData.txt Blocks.txt
+hb-ot-shaper-arabic-table.hh: gen-arabic-table.py ArabicShaping.txt UnicodeData.txt Blocks.txt
 	./$^ > $@ || ($(RM) $@; false)
 hb-unicode-emoji-table.hh: gen-emoji-table.py emoji-data.txt emoji-test.txt
 	./$^ > $@ || ($(RM) $@; false)
-hb-ot-shape-complex-indic-table.cc: gen-indic-table.py IndicSyllabicCategory.txt IndicPositionalCategory.txt Blocks.txt
+hb-ot-shaper-indic-table.cc: gen-indic-table.py IndicSyllabicCategory.txt IndicPositionalCategory.txt Blocks.txt
 	./$^ > $@ || ($(RM) $@; false)
 hb-ot-tag-table.hh: gen-tag-table.py languagetags language-subtag-registry
 	./$^ > $@ || ($(RM) $@; false)
 hb-ucd-table.hh: gen-ucd-table.py ucd.nounihan.grouped.zip hb-common.h
 	./$^ > $@ || ($(RM) $@; false)
-hb-ot-shape-complex-use-table.hh: gen-use-table.py IndicSyllabicCategory.txt IndicPositionalCategory.txt ArabicShaping.txt DerivedCoreProperties.txt UnicodeData.txt Blocks.txt Scripts.txt ms-use/IndicSyllabicCategory-Additional.txt ms-use/IndicPositionalCategory-Additional.txt
+hb-ot-shaper-use-table.hh: gen-use-table.py IndicSyllabicCategory.txt IndicPositionalCategory.txt ArabicShaping.txt DerivedCoreProperties.txt UnicodeData.txt Blocks.txt Scripts.txt ms-use/IndicSyllabicCategory-Additional.txt ms-use/IndicPositionalCategory-Additional.txt
 	./$^ > $@ || ($(RM) $@; false)
-hb-ot-shape-complex-vowel-constraints.cc: gen-vowel-constraints.py ms-use/IndicShapingInvalidCluster.txt Scripts.txt
+hb-ot-shaper-vowel-constraints.cc: gen-vowel-constraints.py ms-use/IndicShapingInvalidCluster.txt Scripts.txt
 	./$^ > $@ || ($(RM) $@; false)
 
 packtab:
